Add integration tests via Cucumber
The trick here is to use scren to spawn a subshell that can run
'sl' without hosing the test terminals stdout. It writes the
retcode to a tempfile so it can be retrieved and used as the
exit code from the 'helper' script.

Also update the docker container to have 'screen' installed.